+++ src/components/spec-document/ArticleNode.vue
