+++ src/components/spec-document/ModelNode.spec.ts
