+++ src/components/spec-document/ModelNode.vue
