+++ src/components/spec-document/ModelProperty.spec.ts
