+++ src/components/spec-document/index.ts
