+++ src/components/spec-document/property-fields/PropertyAnyOf.spec.ts
