+++ src/components/spec-document/property-fields/PropertyAnyOf.vue
