+++ src/components/spec-document/property-fields/PropertyDescription.spec.ts
