+++ src/components/spec-document/property-fields/PropertyDescription.vue
