+++ src/components/spec-document/property-fields/PropertyEnum.spec.ts
