+++ src/components/spec-document/property-fields/PropertyEnum.vue
