+++ src/components/spec-document/property-fields/PropertyExample.spec.ts
