+++ src/components/spec-document/property-fields/PropertyExample.vue
