+++ src/components/spec-document/property-fields/PropertyInfo.spec.ts
