+++ src/components/spec-document/property-fields/PropertyInfo.vue
