+++ src/components/spec-document/property-fields/PropertyPattern.spec.ts
