+++ src/components/spec-document/property-fields/PropertyPattern.vue
