+++ src/components/spec-document/property-fields/PropertyRange.spec.ts
