+++ src/components/spec-document/property-fields/PropertyRange.vue
