+++ src/components/spec-renderer-toc/ExternalLinkItem.vue
