+++ src/components/spec-renderer-toc/GroupItem.vue
