+++ src/components/spec-renderer-toc/NodeGroupItem.vue
